move storage and awsconfig to internal
Just move the code around so it's more best practices
