+++ internal/awsconfig/config.go
@@ -23,8 +23,8 @@ import (
 	//	"os"
 	"strings"
 
+	"github.com/synfinatic/aws-sso-cli/internal/storage"
 	"github.com/synfinatic/aws-sso-cli/internal/utils"
-	"github.com/synfinatic/aws-sso-cli/storage"
 	"gopkg.in/ini.v1"
 )
