+++ internal/storage/testdata/aws-sso-cli-records
