+++ internal/storage/testdata/bad_store.json
